back out pull/export naming change
old "pull" is "pull" again; new command is "checkout"
